fix: Add missing torch import in overlap detector
- Fixed 'name torch is not defined' error in overlap_detector.py
- Added torch import to optional imports section
- Added check for PYANNOTE_AVAILABLE before using torch
- This fixes the error that occurs during audio enhancement with secondary speaker detection

+++ tests/test_append_mode_integration.py
@@ -51,7 +51,7 @@ class TestAppendModeIntegration(unittest.TestCase):
             "transcript": Value("string"),
             "length": Value("float32"),
             "dataset_name": Value("string"),
-            "confidence_score": Value("float64")
+            "confidence_score": Value("float32")
         })
         
         existing_dataset = Dataset.from_dict(existing_data, features=features)
